Attempt #1 at a stable visualization.
Fails for a couple of reasons. First, the cache is not trimmed properly
because the 'delete' timestamps are not the same as the 'insert'
timestamps. Second, updates are POSTed as individual methods which means
that applying them immediately results in an incomplete dataset. The
server needs to buffer messages until it can produce an atomic set.